Revert "Moved extract media function to different file to fix backend functions build"
This reverts commit d2550e0657dc736bfbb73cc7852533b8808a09c8.
