+++ libs/organization/src/lib/organization/pages/organization/organization.component.ts
@@ -4,7 +4,7 @@ import { OrganizationQuery } from '@blockframes/organization/+state/organization
 import { OrganizationService } from '@blockframes/organization/+state/organization.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MediaService } from '@blockframes/media/+state/media.service';
-import { extractMediaFromDocumentBeforeUpdate } from '@blockframes/media/+state/media.utils';
+import { extractMediaFromDocumentBeforeUpdate } from '@blockframes/media/+state/media.model';
 
 @Component({
   selector: 'organization-edit',
@@ -34,7 +34,7 @@ export class OrganizationComponent implements OnInit {
           throw new Error('Your organization profile information are not valid');
         }
 
-        const { documentToUpdate, mediasToUpload } = extractMediaFromDocumentBeforeUpdate(this.organizationForm);
+        const { documentToUpdate, mediasToUpload } = extractMediaFromDocumentBeforeUpdate(this.organizationForm.value, this.organizationForm);
         this.service.update(this.query.getActiveId(), documentToUpdate);
         this.mediaService.uploadOrDeleteMedia(mediasToUpload);
